Update init_spec.rb to cover base nagios::resource
The top level nagios class is the only place we pull in the default
nagios::resource definitions. This updates the spec test for that class
to validate that all of the default resources are being defined, though
not verifying the full value sets as the nagios::resource class tests
validate the data passing far better and the final nagios_* resource
types do their own parameter validation.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>