Switch from using create_resources
There is an issue with recent versions of puppet-agent when being used
under puppetserver that seems to cause exported files to have problems.
It's possible that create_resources is somehow to blame and the new
lambda method may get around this.

In any case, switching to the lambda methodology makes sense for the
long run anyway.

https://tickets.puppetlabs.com/browse/PUP-6698

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>